Support production within msParts
